Support multiple language helper functions (#16)
Moves a lot of stuff around. See README changes and tests for semgrepl functionality.

This shows how support queries for multiple languages. This has most of #2, just need to complete the patterns.

I opted to pass config around as a variable to ease development. If we store configs as global variable, then ipython autoreload clobbers the config after every command.